Fix setup module violation
