+++ frontend/src/metabase/setup/actions.ts
@@ -9,7 +9,7 @@ import {
   updateSetting,
   updateSettings,
 } from "metabase/redux/settings";
-import type { InviteInfo, Locale, State, UserInfo } from "metabase/redux/store";
+import type { InviteInfo, Locale, SetupStep, State, UserInfo  } from "metabase/redux/store";
 import { createAsyncThunk } from "metabase/redux/utils";
 import { getSetting } from "metabase/selectors/settings";
 import { SetupApi } from "metabase/services";
@@ -31,7 +31,6 @@ import {
   getSetupToken,
   getUsageReason,
 } from "./selectors";
-import type { SetupStep } from "./types";
 import { getDefaultLocale, getLocales } from "./utils";
 
 interface ThunkConfig {
