+++ frontend/src/metabase/setup/components/DataUsageStep/DataUsageStep.unit.spec.tsx
@@ -6,11 +6,11 @@ import {
   setupSettingsEndpoints,
 } from "__support__/server-mocks";
 import { renderWithProviders, screen, waitFor } from "__support__/ui";
+import type { SetupStep } from "metabase/redux/store";
 import {
   createMockSetupState,
   createMockState,
 } from "metabase/redux/store/mocks";
-import type { SetupStep } from "metabase/setup/types";
 import { createMockSettings } from "metabase-types/api/mocks";
 
 import { DataUsageStep } from "./DataUsageStep";
