+++ frontend/src/metabase/setup/components/DatabaseStep/DatabaseStep.unit.spec.tsx
@@ -2,12 +2,12 @@ import userEvent from "@testing-library/user-event";
 import fetchMock from "fetch-mock";
 
 import { renderWithProviders, screen, within } from "__support__/ui";
+import type { SetupStep } from "metabase/redux/store";
 import {
   createMockSettingsState,
   createMockSetupState,
   createMockState,
 } from "metabase/redux/store/mocks";
-import type { SetupStep } from "metabase/setup/types";
 import type { DatabaseData } from "metabase-types/api";
 import { createMockDatabaseData } from "metabase-types/api/mocks";
