+++ frontend/src/metabase/setup/components/SettingsPage/SettingsPage.tsx
@@ -1,8 +1,8 @@
 import { getLocalizationNoticeText } from "metabase/common/components/CommunityLocalizationNotice";
 import { LogoIcon } from "metabase/common/components/LogoIcon";
 import { useSelector } from "metabase/redux";
+import type { SetupStep } from "metabase/redux/store";
 import { getSteps } from "metabase/setup/selectors";
-import type { SetupStep } from "metabase/setup/types";
 import { Box, Flex, Icon, Text, Tooltip } from "metabase/ui";
 
 import { CloudMigrationHelp } from "../CloudMigrationHelp";
