+++ frontend/src/metabase/setup/components/UserStep/UserStep.unit.spec.tsx
@@ -1,12 +1,11 @@
 import { mockSettings } from "__support__/settings";
 import { renderWithProviders, screen } from "__support__/ui";
-import type { UserInfo } from "metabase/redux/store";
+import type { SetupStep, UserInfo  } from "metabase/redux/store";
 import {
   createMockSetupState,
   createMockState,
   createMockUserInfo,
 } from "metabase/redux/store/mocks";
-import type { SetupStep } from "metabase/setup/types";
 
 import { UserStep } from "./UserStep";
