--- frontend/src/metabase/setup/types.ts
@@ -1,9 +1,0 @@
-export type SetupStep =
-  | "welcome"
-  | "language"
-  | "user_info"
-  | "usage_question"
-  | "db_connection"
-  | "license_token"
-  | "data_usage"
-  | "completed";
